Improve completions
